requirements.txt
anthropic>=0.5.0
Pillow>=9.0.0
rich>=10.0.0
pytest>=7.0.0
pytest-cov>=2.0.0
